Analyzer no longer double counts
This update prevents the MIDI Analyzer from double counting notes when present in a chord. In addition, it ignores notes under the BRE and no longer counts the notes after the BRE in the total note count (they are now also separated in the analysis).

I've also condensed guitar, bass, and 5-lane keys to only use one function instead of 3 separate, but similar, ones.

This only applies to guitar, bass, and 5-lane keys for now. Drums and Pro Keys are a bit more complicated.